chore(html): rename i18n provider module
