+++ packages/html/src/i18n/index.ts
@@ -26,4 +26,4 @@ export { createI18n } from './create-i18n';
 export { MediaI18nProviderElement, MediaTextElement } from './elements';
 export { context, I18nController, I18nTextMixin } from './instance';
 export { localeFromDomLang, resolvePlayerLocale, resolveProviderLocale } from './locale';
-export { I18nProviderMixin } from './provider-instance';
+export { I18nProviderMixin } from './provider';
